Move hasBeenSentRecently
