+++ lib/publication/index.js
@@ -8,13 +8,20 @@ const w = require('../util/w')
 const {client, dbName} = require('../util/mongo')
 const fcStrategy = require('./auth/franceconnect')
 const {getSubmission, updateSubmission, createSubmission, getPublishedSubmissions} = require('./models')
-const {generatePinCode, getExpirationDate, validatePinCode, hasBeenSentRecently} = require('../util/pin-code')
+const {generatePinCode, getExpirationDate, validatePinCode} = require('../util/pin-code')
 const {sendMail} = require('../util/sendmail')
 const createPinCodeEmail = require('../emails/pin-code')
 
 const app = new Router()
 
 app.use(json())
+
+function hasBeenSentRecently(date) {
+  const now = new Date()
+  const coolDownTime = new Date(date)
+  coolDownTime.setMinutes(coolDownTime.getMinutes() + 5)
+  return now > coolDownTime
+}
 
 const sessionOptions = {
   secret: process.env.SESSION_SECRET || 'foobar',
