+++ lib/util/pin-code.js
@@ -15,11 +15,4 @@ function validatePinCode(codeA, codeB, expirationDate) {
   return codeA === codeB && now < expirationDate
 }
 
-function hasBeenSentRecently(date) {
-  const now = new Date()
-  const coolDownTime = new Date(date)
-  coolDownTime.setMinutes(coolDownTime.getMinutes() + 5)
-  return now > coolDownTime
-}
-
-module.exports = {generatePinCode, getExpirationDate, validatePinCode, hasBeenSentRecently}
+module.exports = {generatePinCode, getExpirationDate, validatePinCode}
